+++ backend/blog-backend/src/main/java/com/kisesaki/blog/auth/dto/auth/request/ForgotPasswordRequestDto.java
@@ -1,4 +1,4 @@
-package com.kisesaki.blog.auth.dto.request;
+package com.kisesaki.blog.auth.dto.auth.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
