+++ backend/blog-backend/src/main/java/com/kisesaki/blog/auth/dto/auth/request/LogoutRequestDto.java
@@ -1,4 +1,4 @@
-package com.kisesaki.blog.auth.dto.request;
+package com.kisesaki.blog.auth.dto.auth.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
