+++ backend/blog-backend/src/main/java/com/kisesaki/blog/auth/dto/auth/request/ResetPasswordRequestDto.java
@@ -1,4 +1,4 @@
-package com.kisesaki.blog.auth.dto.request;
+package com.kisesaki.blog.auth.dto.auth.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
@@ -7,9 +7,9 @@ import lombok.Data;
 
 @Data
 @AllArgsConstructor
-public class ChangePasswordRequestDto {
-    @NotBlank(message = "旧密码不能为空")
-    private String oldPassword;
+public class ResetPasswordRequestDto {
+    @NotBlank(message = "邮箱验证令牌不能为空")
+    private String resetToken;
 
     @NotBlank(message = "新密码不能为空")
     @Size(min = 6, max = 100, message = "密码长度必须在6-100字符之间")
