+++ backend/blog-backend/src/main/java/com/kisesaki/blog/auth/dto/auth/response/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.kisesaki.blog.auth.dto.response;
+package com.kisesaki.blog.auth.dto.auth.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
